requirements.txt
regex
requests
discord.py
discord-py-slash-command